+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # HEAD
 
+<<<<<<< HEAD
 -   Nicer shell outputs, with colored errors, warnings and successes [#11](https://github.com/ChiccoDorato/RayCharles/pull/11).
 -   **Breaking change**: added a DSL which parses txt files. Command `demo` renamed to `render`. Improved rendering with implementation of new shapes (CylinderShell and Cylinder) and antialiasing [#10](https://github.com/ChiccoDorato/RayCharles/pull/10).
 
 # Version 0.2.0
 
 -   Added a flat renderer, an on-off renderer and a path tracer. `demo` produces a new image [#9](https://github.com/ChiccoDorato/RayCharles/pull/9). 
+=======
+-   Added flat renderer and on-off renderer [#9](https://github.com/ChiccoDorato/RayCharles/pull/9). 
+>>>>>>> origin/pathtracing
 -   **Breaking change**: new CLI which provides help and supports the commands `pfm2png` and `demo` [#6](https://github.com/ChiccoDorato/RayCharles/pull/6).
 -   Fix an issue with the vertical order of the images [#5](https://github.com/ChiccoDorato/RayCharles/pull/5).